+++ integration/astro7/.storybook/vitest.setup.ts
@@ -1,0 +1,7 @@
+import '../lib/vitest-setup';
+import { setProjectAnnotations } from '@storybook-astro/framework';
+import * as projectAnnotations from './preview';
+
+// This is an important step to apply the right configuration when testing your stories.
+// More info at: https://storybook.js.org/docs/api/portable-stories/portable-stories-vitest#setprojectannotations
+setProjectAnnotations([projectAnnotations]);